refactor(react-langchain): share the langchain converter layer with react-langgraph
