+++ .changeset/langchain-converter-seam.md
@@ -1,0 +1,6 @@
+---
+"@assistant-ui/react-langchain": patch
+"@assistant-ui/react-langgraph": patch
+---
+
+refactor: share the langchain converter layer with react-langgraph. @langchain/react becomes an optional peer of react-langchain.
