+++ packages/react-langgraph/package.json
@@ -35,6 +35,7 @@
   },
   "dependencies": {
     "@assistant-ui/core": "^0.3.15",
+    "@assistant-ui/react-langchain": "^0.0.27",
     "@assistant-ui/store": "^0.3.10",
     "assistant-cloud": "*",
     "assistant-stream": "^0.3.39"
