+++ packages/react-langgraph/src/convertLangChainMessages.ts
@@ -1,7 +1,6 @@
 "use client";
 
 import type {
-  AppendMessage,
   CompleteAttachment,
   DataMessagePart,
   MessageTiming,
@@ -11,11 +10,16 @@ import type {
 } from "@assistant-ui/core";
 import type { useExternalMessageConverter } from "@assistant-ui/core/react";
 import {
-  httpUrlPattern,
   parseDataUrl,
   stableStringifyToolArgs,
   trackToolArgsKeyOrder,
 } from "@assistant-ui/core/internal";
+import {
+  convertLangChainContentBlock,
+  getCustomMetadata,
+  uiMessageToDataPart,
+  withAudioTranscript,
+} from "@assistant-ui/react-langchain/converter";
 import type {
   LangChainMessage,
   LangChainToolCall,
@@ -34,12 +38,6 @@ type LangGraphMessageConverterMetadata =
     messageTiming?: Record<string, MessageTiming>;
     attachmentsByMessageId?: Map<string, readonly CompleteAttachment[]>;
   };
-
-const uiMessageToDataPart = (ui: UIMessage): DataMessagePart => ({
-  type: "data",
-  name: ui.name,
-  data: ui.props,
-});
 
 const getToolArgsCacheKey = (
   messageId: string | undefined,
@@ -92,11 +90,6 @@ const resolveToolCallArgs = ({
   return { args, argsText };
 };
 
-const getCustomMetadata = (
-  additionalKwargs: Record<string, unknown> | undefined,
-): Record<string, unknown> =>
-  (additionalKwargs?.metadata as Record<string, unknown>) ?? {};
-
 const warnedMessagePartTypes = new Set<string>();
 const warnForUnknownMessagePartType = (type: string) => {
   if (
@@ -136,90 +129,27 @@ const contentToParts = (
       ):
         | (ThreadUserMessage | ThreadAssistantMessage)["content"][number]
         | null => {
-        const type = part.type;
-        switch (type) {
-          case "text":
-            return { type: "text", text: part.text };
-          case "text_delta":
-            return { type: "text", text: part.text };
-          case "image_url": {
-            const image =
-              typeof part.image_url === "string"
-                ? part.image_url
-                : part.image_url?.url;
-            if (!image) return null;
-            return { type: "image", image };
-          }
-          case "file":
-            return {
-              type: "file",
-              filename: part.metadata?.filename ?? "file",
-              data:
-                part.source_type === "url"
-                  ? part.url
-                  : part.source_type === "id"
-                    ? part.id
-                    : part.data,
-              mimeType: part.mime_type ?? "application/octet-stream",
-              ...((part.source_type === "url" || part.source_type === "id") && {
-                sourceType: part.source_type,
-              }),
-            };
-
-          case "audio": {
-            const mimeType = part.mime_type ?? "application/octet-stream";
-            const subtype = mimeType.startsWith("audio/")
-              ? mimeType.slice("audio/".length)
-              : undefined;
-            return {
-              type: "file" as const,
-              filename: subtype ? `audio.${subtype}` : "audio",
-              data: part.data,
-              mimeType,
-            };
-          }
-
-          case "thinking":
-            return { type: "reasoning", text: part.thinking };
-
-          case "reasoning":
-            return {
-              type: "reasoning",
-              text:
-                part.summary?.map((s) => s?.text ?? "").join("\n\n\n") ??
-                part.reasoning ??
-                "",
-            };
-
-          case "tool_use":
-            return null;
-          case "input_json_delta":
-            return null;
-
-          case "computer_call": {
-            const args = part.action as ReadonlyJSONObject;
-            return {
-              type: "tool-call",
-              toolCallId: part.call_id,
-              toolName: "computer_call",
+        if (part.type === "computer_call") {
+          const args = part.action as ReadonlyJSONObject;
+          return {
+            type: "tool-call",
+            toolCallId: part.call_id,
+            toolName: "computer_call",
+            args,
+            argsText: stableStringifyToolArgs(
+              metadata.toolArgsKeyOrderCache,
+              getToolArgsCacheKey(messageId, "computer", part.call_id),
               args,
-              argsText: stableStringifyToolArgs(
-                metadata.toolArgsKeyOrderCache,
-                getToolArgsCacheKey(messageId, "computer", part.call_id),
-                args,
-              ),
-            };
-          }
-
-          default: {
-            const _exhaustiveCheck: never = type;
-            warnForUnknownMessagePartType(_exhaustiveCheck);
-            return null;
-          }
-
-          // const _exhaustiveCheck: never = type;
-          // throw new Error(`Unknown message part type: ${_exhaustiveCheck}`);
+            ),
+          };
         }
+
+        const converted = convertLangChainContentBlock(part);
+        if (converted === undefined) {
+          warnForUnknownMessagePartType(part.type);
+          return null;
+        }
+        return converted;
       },
     )
     .filter((a) => a !== null);
@@ -272,33 +202,6 @@ const dropAttachmentDuplicates = (
   }
   if (dropped.size === 0) return parts;
   return parts.filter((_, i) => !dropped.has(i));
-};
-
-const hasVisibleText = (text: unknown): boolean =>
-  typeof text === "string" && text.trim() !== "";
-
-/**
- * Audio output arrives outside the content array: providers leave `content`
- * empty and put the spoken text in `additional_kwargs.audio.transcript`. The
- * audio bytes stay behind because no provider reports their media type, and a
- * streamed response carries raw PCM rather than a playable file.
- */
-const withAudioTranscript = (
-  parts: ReturnType<typeof contentToParts>,
-  additionalKwargs: Extract<
-    LangChainMessage,
-    { type: "ai" }
-  >["additional_kwargs"],
-): ReturnType<typeof contentToParts> => {
-  const transcript: unknown = additionalKwargs?.audio?.transcript;
-  if (typeof transcript !== "string" || !hasVisibleText(transcript))
-    return parts;
-  if (parts.some((part) => part.type === "text" && hasVisibleText(part.text)))
-    return parts;
-  return [
-    ...parts.filter((part) => part.type !== "text"),
-    { type: "text" as const, text: transcript },
-  ];
 };
 
 export const convertLangChainMessages: useExternalMessageConverter.Callback<
@@ -414,114 +317,4 @@ export const convertLangChainMessages: useExternalMessageConverter.Callback<
   }
 };
 
-/**
- * Audio media types that reach a provider's audio input through the LangChain
- * `audio` block. langchain-core derives OpenAI's `input_audio.format` by
- * splitting `mime_type` on `/`, and that format is a wav-or-mp3 enum, so
- * `audio/mpeg` passes the converter and is rejected at the provider.
- */
-const audioBlockMimeTypes = new Map<string, "audio/mp3" | "audio/wav">([
-  ["audio/mp3", "audio/mp3"],
-  ["audio/mpeg", "audio/mp3"],
-  ["audio/wav", "audio/wav"],
-  ["audio/wave", "audio/wav"],
-  ["audio/x-wav", "audio/wav"],
-]);
-
-export const getMessageContent = (msg: AppendMessage) => {
-  const allContent = [
-    ...msg.content,
-    ...(msg.attachments?.flatMap((a) => a.content) ?? []),
-  ];
-
-  const hasNonText = allContent.some(
-    (part) =>
-      part.type === "file" || part.type === "image" || part.type === "audio",
-  );
-  const hasText = allContent.some((part) => part.type === "text");
-  if (hasNonText && !hasText) {
-    allContent.unshift({ type: "text", text: " " });
-  }
-
-  const content = allContent.flatMap((part) => {
-    const type = part.type;
-    switch (type) {
-      case "text":
-        return { type: "text" as const, text: part.text };
-      case "image":
-        return { type: "image_url" as const, image_url: { url: part.image } };
-      case "file": {
-        const metadata = { filename: part.filename ?? "file" };
-        if (part.sourceType === "id") {
-          return {
-            type: "file" as const,
-            id: part.data,
-            mime_type: part.mimeType,
-            filename: metadata.filename,
-            metadata,
-            source_type: "id" as const,
-          };
-        }
-        if (part.sourceType === "url" || httpUrlPattern.test(part.data)) {
-          return {
-            type: "file" as const,
-            url: part.data,
-            mime_type: part.mimeType,
-            filename: metadata.filename,
-            metadata,
-            source_type: "url" as const,
-          };
-        }
-        const parsed = parseDataUrl(part.data);
-        const audioMimeType = audioBlockMimeTypes.get(
-          (parsed?.mimeType ?? part.mimeType).toLowerCase(),
-        );
-        if (audioMimeType) {
-          return {
-            type: "audio" as const,
-            data: parsed?.data ?? part.data,
-            mime_type: audioMimeType,
-            source_type: "base64" as const,
-          };
-        }
-        return {
-          type: "file" as const,
-          data: parsed?.data ?? part.data,
-          mime_type: parsed?.mimeType ?? part.mimeType,
-          filename: metadata.filename,
-          metadata,
-          source_type: "base64" as const,
-        };
-      }
-
-      case "audio": {
-        const parsed = parseDataUrl(part.audio.data);
-        return {
-          type: "audio" as const,
-          data: parsed?.data ?? part.audio.data,
-          mime_type: `audio/${part.audio.format}`,
-          source_type: "base64" as const,
-        };
-      }
-
-      case "data":
-        return [];
-
-      case "tool-call":
-        throw new Error("Tool call appends are not supported.");
-
-      default: {
-        const _exhaustiveCheck: "reasoning" | "source" | "generative-ui" = type;
-        throw new Error(
-          `Unsupported append message part type: ${_exhaustiveCheck}`,
-        );
-      }
-    }
-  });
-
-  if (content.length === 1 && content[0]?.type === "text") {
-    return content[0].text ?? "";
-  }
-
-  return content;
-};
+export { getMessageContent } from "@assistant-ui/react-langchain/converter";
